pyproject.toml
[tool.poetry]
name = "gmail-api"
version = "0.2.0"
description = ""
authors = ["bradi <bradisbon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.12.3"
httplib2shim = "^0.0.3"
oauth2client = "^4.1.3"
google-auth-oauthlib = "^0.4.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
